Move all the http abstractions into a single pusher.http module
