+++ pusher/util.py
@@ -3,8 +3,6 @@
 from __future__ import (print_function, unicode_literals, absolute_import,
                         division)
 
-from pusher.errors import *
-
 import json
 import re
 import six
@@ -13,24 +11,10 @@ import sys
 channel_name_re = re.compile('^[-a-zA-Z0-9_=@,.;]+$')
 app_id_re       = re.compile('^[0-9]+$')
 
-GET, POST, PUT, DELETE = "GET", "POST", "PUT", "DELETE"
-
 if sys.version_info < (3,):
     text = 'a unicode string'
 else:
     text = 'a string'
-
-def process_response(status, body):
-    if status == 200:
-        return json.loads(body)
-    elif status == 400:
-        raise PusherBadRequest(body)
-    elif status == 401:
-        raise PusherBadAuth(body)
-    elif status == 403:
-        raise PusherForbidden(body)
-    else:
-        raise PusherBadStatus("%s: %s" % (status, body))
 
 def validate_channel(channel):
     if not isinstance(channel, six.text_type):
